pyproject.toml
[tool.poetry]
name = "fastapi-auth"
version = "0.2.0"
description = "A robust authentication and user management system built with FastAPI and SQLModel"
authors = ["DocSJ1980 <docsaj@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
sqlmodel = "^0.0.21"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pytest = "^8.3.2"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]